server/app/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Settings:
    MONGO_URI = os.getenv("MONGO_URI")
    DB_NAME = os.getenv("DB_NAME")
    JWT_SECRET = os.getenv("JWT_SECRET")
    JWT_EXPIRE_HOURS = int(os.getenv("JWT_EXPIRE_HOURS", 8))

settings = Settings()
