server/requirements.txt
fastapi
uvicorn
motor
python-dotenv
python-jose
passlib[bcrypt]
pydantic
email-validator
